+++ src/main/java/com/besson/endfield/ArknightsEndfieldClient.java
@@ -4,6 +4,7 @@ import com.besson.endfield.blockentity.ModBlockEntities;
 import com.besson.endfield.renderer.block.*;
 import com.besson.endfield.screen.ModScreens;
 import com.besson.endfield.screen.custom.PortableOriginiumRigScreen;
+import com.besson.endfield.screen.custom.ProtocolAnchorCoreScreen;
 import net.fabricmc.api.ClientModInitializer;
 import net.minecraft.client.gui.screen.ingame.HandledScreens;
 import net.minecraft.client.render.block.entity.BlockEntityRendererFactories;
@@ -31,5 +32,6 @@ public class ArknightsEndfieldClient implements ClientModInitializer {
         BlockEntityRendererFactories.register(ModBlockEntities.THERMAL_BANK, ThermalBankRenderer::new);
 
         HandledScreens.register(ModScreens.PORTABLE_ORIGINIUM_RIG_SCREEN, PortableOriginiumRigScreen::new);
+        HandledScreens.register(ModScreens.PROTOCOL_ANCHOR_CORE_SCREEN, ProtocolAnchorCoreScreen::new);
     }
 }